# app/src/main/java/org/lfenergy/compas/scl/data/rest/api/locations/model/Location.java
// SPDX-FileCopyrightText: 2024 BearingPoint GmbH
//
// SPDX-License-Identifier: Apache-2.0
package org.lfenergy.compas.scl.data.rest.api.locations.model;

import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonTypeName;
import jakarta.validation.constraints.NotNull;

import java.util.Objects;



@JsonTypeName("Location")
@jakarta.annotation.Generated(value = "org.openapitools.codegen.languages.JavaJAXRSSpecServerCodegen", date = "2024-11-18T07:52:46.875467800+01:00[Europe/Vienna]", comments = "Generator version: 7.8.0")
public class Location   {
  private String uuid;
  private String key;
  private String name;
  private String description;
  private Integer assignedResources;

  /**
   * Unique location uuid generated by backend during creation
   **/
  public Location uuid(String uuid) {
    this.uuid = uuid;
    return this;
  }


  @JsonProperty("uuid")
  public String getUuid() {
    return uuid;
  }

  @JsonProperty("uuid")
  public void setUuid(String uuid) {
    this.uuid = uuid;
  }

  /**
   * Location key, defined once manually when creating a location
   **/
  public Location key(String key) {
    this.key = key;
    return this;
  }


  @JsonProperty("key")
  @NotNull public String getKey() {
    return key;
  }

  @JsonProperty("key")
  public void setKey(String key) {
    this.key = key;
  }

  /**
   * Location name
   **/
  public Location name(String name) {
    this.name = name;
    return this;
  }


  @JsonProperty("name")
  @NotNull public String getName() {
    return name;
  }

  @JsonProperty("name")
  public void setName(String name) {
    this.name = name;
  }

  /**
   * Location description
   **/
  public Location description(String description) {
    this.description = description;
    return this;
  }


  @JsonProperty("description")
  public String getDescription() {
    return description;
  }

  @JsonProperty("description")
  public void setDescription(String description) {
    this.description = description;
  }

  /**
   * Number of resources assigned to this location
   **/
  public Location assignedResources(Integer assignedResources) {
    this.assignedResources = assignedResources;
    return this;
  }


  @JsonProperty("assignedResources")
  public Integer getAssignedResources() {
    return assignedResources;
  }

  @JsonProperty("assignedResources")
  public void setAssignedResources(Integer assignedResources) {
    this.assignedResources = assignedResources;
  }


  @Override
  public boolean equals(Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    Location location = (Location) o;
    return Objects.equals(this.uuid, location.uuid) &&
        Objects.equals(this.key, location.key) &&
        Objects.equals(this.name, location.name) &&
        Objects.equals(this.description, location.description) &&
        Objects.equals(this.assignedResources, location.assignedResources);
  }

  @Override
  public int hashCode() {
    return Objects.hash(uuid, key, name, description, assignedResources);
  }

  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class Location {\n");

    sb.append("    uuid: ").append(toIndentedString(uuid)).append("\n");
    sb.append("    key: ").append(toIndentedString(key)).append("\n");
    sb.append("    name: ").append(toIndentedString(name)).append("\n");
    sb.append("    description: ").append(toIndentedString(description)).append("\n");
    sb.append("    assignedResources: ").append(toIndentedString(assignedResources)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }


}